chore: Move ArticleTitleInput to ArticleEditor directory
